% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonlightModel.R
\name{elevExtCoeff}
\alias{elevExtCoeff}
\title{Calculate extinction coefficient based on elevation of the observer}
\usage{
elevExtCoeff(elev)
}
\arguments{
\item{elev}{elevation in meters asl}
}
\value{
A single numeric value: the estimated extinction coefficient for the
  given elevation.
}
\description{
Calculate extinction coefficient based on elevation of the observer
}
\examples{
# Extinction coefficient at sea level
suppressWarnings(elevExtCoeff(0))

# Extinction coefficient at 1000 m asl (e.g. approximate for many study sites)
suppressWarnings(elevExtCoeff(1000))
}
