% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonlightModel.R
\name{calculateMoonlightIntensity}
\alias{calculateMoonlightIntensity}
\title{Calculate moonlight intensity}
\usage{
calculateMoonlightIntensity(lat, lon, date, e)
}
\arguments{
\item{e}{}
}
\value{

}
\description{
This function is predicting moonlight intensity on the ground for any given place and time,
based on the location, position of the moon and number of correction factors
}
\examples{






}
