% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonlightModel.R
\name{calculateMoonlightIntensity}
\alias{calculateMoonlightIntensity}
\title{Calculate moonlight intensity on the ground}
\usage{
calculateMoonlightIntensity(lat, lon, date, e)
}
\arguments{
\item{e}{}
}
\value{

}
\description{
Calculate moonlight intensity on the ground
}
