% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMoonlightValues.R
\name{calculateMoonlightStatistics}
\alias{calculateMoonlightStatistics}
\title{Title}
\usage{
calculateMoonlightStatistics(lat, lon, date, e, t, timezone)
}
\arguments{
\item{lat}{latitude, numerical decimal}

\item{lon}{longitude, numerical decimal}

\item{date}{date time as POSIXct with the local time zone. If needed use
as.POSIXct(date, tz=timezone)}

\item{e}{extinction coefficient - a single numerical value depending on the
altitude. Average extinction coefficients (magnitude per air mass) are as
follows: (At sea level: 0.28; at 500m asl: 0.24; at 1000m asl: 0.21; at
2000m asl: 0.16)}

\item{t}{sampling interval. It is used in seq() function so the same values
are accepted: A character string, containing one of "sec", "min", "hour".
This can optionally be preceded by a (positive or negative) integer and a
space, or followed by "s". Example: "15 mins", "3 hour" etc.}

\item{timezone}{time zone of the data usually in the format "Continent/City",
i.e. for Poland: "Europe/Warsaw"}
}
\value{
output
}
\description{
Title
}
