% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMoonlightValues.R
\name{calculateMoonlightStatistics}
\alias{calculateMoonlightStatistics}
\title{Title}
\usage{
calculateMoonlightStatistics(lat, lon, date, e, t, timezone)
}
\arguments{
\item{timezone}{}
}
\value{
data
}
\description{
Title
}
