% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonlightModel.R
\name{elevExtCoeff}
\alias{elevExtCoeff}
\title{Calculate extinction coefficient based on elevation of the observer}
\usage{
elevExtCoeff(elev)
}
\arguments{
\item{elev}{elevation in meters asl}
}
\value{
elevExtCoeff
}
\description{
Calculate extinction coefficient based on elevation of the observer
}
\examples{
\dontrun{

elev <- 2228 # Elevation in m asl for Mount Kosciuszko
result <- elevExtCoeff(elev)
}

}
